Compile in gnu++XX mode to get correct 128-bit integer limits/traits
https://quuxplusone.github.io/blog/2019/02/28/is-int128-integral/
"""
If you compile in -std=gnu++XX mode, libstdc++ will treat __int128 and
unsigned __int128 as integral types with the proper numeric limits; but
if you compile in the usual -std=c++XX mode, libstdc++ will not do any
of that.

On the other hand, libc++ treats __int128 as an integer type, full stop.
"""