Refactor the reading of input
Instead of reading line-by-line from sys.stdin, read as many bytes as
possible from sys.stdin.buffer with readinto1.
If the bytes can be decoded and they don't end with an incomplete escape
sequence, decode the bytes and operate line-by-line.
Else, read more bytes until the above conditions are satisfied.
This follows the behavior of lolcat.